queryを別ファイルに切り出し
