+++ backend/src/schema/queries/index.ts
@@ -1,0 +1,17 @@
+import { myShoppingList } from './myShoppingList.js';
+import { recipe } from './recipe.js';
+import { recipes } from './recipes.js';
+import { recipesByUserId } from './recipesByUserId.js';
+import { searchIngredients } from './searchIngredients.js';
+import { user } from './user.js';
+import { users } from './users.js';
+
+export const Query = {
+  user,
+  users,
+  recipe,
+  recipes,
+  recipesByUserId,
+  searchIngredients,
+  myShoppingList,
+};
