+++ backend/src/schema/queries/recipe.ts
@@ -1,0 +1,11 @@
+import { PrismaClient } from '@prisma/client';
+
+const prisma = new PrismaClient();
+
+// 特定のユーザーを取得するためのリゾルバ
+export const recipe = async (_: unknown, { id }: { id: string }) => {
+  return prisma.recipe.findUnique({
+    where: { id: Number.parseInt(id) },
+    include: { user: true },
+  });
+};
