+++ backend/src/schema/queries/users.ts
@@ -1,0 +1,10 @@
+import { PrismaClient } from '@prisma/client';
+
+const prisma = new PrismaClient();
+
+// 全てのユーザーを取得するためのリゾルバ
+export const users = async () => {
+  return prisma.user.findMany({
+    include: { recipes: true },
+  });
+};
