+++ backend/src/schema/resolver.ts
@@ -1,13 +1,13 @@
-import { DateTimeResolver } from "graphql-scalars";
-import { Mutation } from "./mutation/mutation.js";
-import { Query } from "./query/query.js";
-import { Recipe } from "./resolvers/Recipe.js";
-import { User } from "./resolvers/User.js";
+import { DateTimeResolver } from 'graphql-scalars';
+import { Mutation } from './mutation/mutation.js';
+import { Recipe } from './resolvers/Recipe.js';
+import { User } from './resolvers/User.js';
+import { Query } from './queries/index.js';
 
 export const resolvers = {
-	DateTime: DateTimeResolver,
-	Query,
-	Mutation,
-	User,
-	Recipe,
+  DateTime: DateTimeResolver,
+  Query,
+  Mutation,
+  User,
+  Recipe,
 };
